Rework IBinding
