+++ OpenTabletDriver.Console/Program.cs
@@ -62,9 +62,9 @@ namespace OpenTabletDriver.Console
             CreateCommand<float, float, float, float>(SetDisplayArea, "Sets the display area"),
             CreateCommand<float, float, float, float, float>(SetTabletArea, "Sets the tablet area"),
             CreateCommand<float, float, float>(SetSensitivity, "Sets the relative sensitivity"),
-            CreateCommand<string, string, float>(SetTipBinding, "Sets the current tip binding"),
-            CreateCommand<string, string, int>(SetPenBinding, "Sets the current pen button bindings"),
-            CreateCommand<string, string, int>(SetAuxBinding, "Sets the current express key bindings"),
+            CreateCommand<string, float>(SetTipBinding, "Sets the current tip binding"),
+            CreateCommand<string, int>(SetPenBinding, "Sets the current pen button bindings"),
+            CreateCommand<string, int>(SetAuxBinding, "Sets the current express key bindings"),
             CreateCommand<int>(SetResetTime, "Sets the reset time in milliseconds"),
             CreateCommand<bool>(SetAutoHook, "Sets whether the driver should automatically enable on start"),
             CreateCommand<bool>(SetEnableClipping, "Sets whether inputs should be limited to the specified areas"),
