+++ OpenTabletDriver.Desktop/Binding/KeyBinding.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Linq;
 using OpenTabletDriver.Desktop.Interop;
 using OpenTabletDriver.Desktop.Interop.Input.Keyboard;
@@ -6,39 +5,42 @@ using OpenTabletDriver.Plugin;
 using OpenTabletDriver.Plugin.Attributes;
 using OpenTabletDriver.Plugin.DependencyInjection;
 using OpenTabletDriver.Plugin.Platform.Keyboard;
+using OpenTabletDriver.Plugin.Tablet;
 
 namespace OpenTabletDriver.Desktop.Binding
 {
-    [PluginName("Key Binding")]
-    public class KeyBinding : IBinding, IValidateBinding
+    [PluginName(PLUGIN_NAME)]
+    public class KeyBinding : IBinding
     {
+        private const string PLUGIN_NAME = "Key Binding";
+
         [Resolved]
         public IVirtualKeyboard Keyboard { set; get; }
 
-        [Property("Property")]
-        public string Property { set; get; }
+        [Property("Keys"), PropertyValidated(nameof(ValidKeys))]
+        public string Keys { set; get; }
 
-        public Action Press
+        public void Press(IDeviceReport report)
         {
-            get => () => Keyboard.Press(Property);
+            Keyboard.Press(Keys);
         }
 
-        public Action Release
+        public void Release(IDeviceReport report)
         {
-            get => () => Keyboard.Release(Property);
+            Keyboard.Release(Keys);
         }
 
-        public string[] ValidProperties
+        public static string[] ValidKeys
         {
             get => DesktopInterop.CurrentPlatform switch
             {
                 PluginPlatform.Windows => WindowsVirtualKeyboard.EtoKeysymToVK.Keys.ToArray(),
                 PluginPlatform.Linux   => EvdevVirtualKeyboard.EtoKeysymToEventCode.Keys.ToArray(),
                 PluginPlatform.MacOS   => MacOSVirtualKeyboard.EtoKeysymToVK.Keys.ToArray(),
-                _                       => null
+                _                      => null
             };
         }
 
-        public override string ToString() => BindingTools.GetShortBindingString(this);
+        public override string ToString() => $"{PLUGIN_NAME}: {Keys}";
     }
 }
