--- OpenTabletDriver.Plugin/IValidateBinding.cs
@@ -1,7 +1,0 @@
-namespace OpenTabletDriver.Plugin
-{
-    public interface IValidateBinding
-    {
-        string[] ValidProperties { get; }
-    }
-}
